refactor: 패키지 이름을 ui => controller, application => service로 수정
