+++ src/main/java/woowacourse/auth/config/AuthenticationPrincipalConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import woowacourse.auth.support.JwtTokenProvider;
-import woowacourse.auth.ui.AuthenticationPrincipalArgumentResolver;
+import woowacourse.auth.controller.AuthenticationPrincipalArgumentResolver;
 import woowacourse.shoppingcart.dao.CustomerDao;
 
 @Configuration
